requirements.txt
# Required Python packages get listed here, one per line.
runpod~=1.7.9
dotenv==0.9.9
numpy==2.2.4
opencv-python==4.11.0.86
orshot==0.1.0
pillow==10.4.0
python-dotenv==1.1.0
requests==2.32.3
yt-dlp==2025.3.31